rest files

+++ backend/routes/cart.js
@@ -92,9 +92,9 @@ router.post('/pay', async (req, res) => {
         return sum;
       };
       if (item.length > 1) {
-        return addPrice(parseInt(item[i].albumprice));
+        return addPrice(parseFloat(item[i].albumprice));
       } else {
-        return addPrice(parseInt(item[0].songprice));
+        return addPrice(parseFloat(item[0].songprice));
       }
     });
